Distinguish between outputs and variables
GitHub has action *outputs* (which are wired up to symbols explicitly by the invoker) as well as setting variables which effectively become env vars for the rest of the build.
Azure Pipelines only has the latter. It has no task outputs.

But the two concepts were conflated before now, with the one API we offered leading AzP to set a variable and GitHub to set a task output.
Now we have one API that sets variables in both, and a set output API that only works in GitHub (and no-ops in AzP).

Fixes #8